Added routes to reservations edit and destroy

+++ app/controllers/reservations_controller.rb
@@ -1,12 +1,12 @@
 class ReservationsController < ApplicationController
-    before_action :set_reservation, only: [:destroy]
+    before_action :set_reservation, only:[:show, :edit, :destroy]
     
     def index
         render json: Reservation.all
     end 
 
     def show
-        reservation = set_reservation
+        # reservation = set_reservation
         if reservation
             render json: reservation
         else
@@ -22,10 +22,26 @@ class ReservationsController < ApplicationController
         end 
     end 
 
-    def destroy
-        reservation.destroy
+    def update
+        # reservation = set_reservation
+        if reservation
+            reservation.update(reservation_params)
+            render json: {reservation: reservation, message: "Updated Successfully"}
+        else
+            render json:{error: "Reservation could not be edited."}
+        end 
     end 
 
+    def destroy
+        # reservation = set_reservation
+        if reservation
+            reservation.destroy
+            render json: {reservation: reservation, message: "Your resevation was successfully deleted."}
+        else
+            render json:{error: "Reservation could not be edited."}
+        end 
+    end
+ 
     private
 
     def set_reservation
@@ -35,4 +51,5 @@ class ReservationsController < ApplicationController
     def reservation_params
         params.require(:reservation).permit(:restaurant_name, :booked_time, :booked_email, :notes, :reserved, :user_id)
     end 
+
 end
